+++ examples/docs-sdk-quickstart/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @executor-js/example-docs-sdk-quickstart
 
+## 0.0.36
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @executor-js/sdk@1.5.33
+  - @executor-js/plugin-openapi@1.5.33
+
 ## 0.0.35
 
 ### Patch Changes