+++ packages/kernel/runtime-workerd-subprocess/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @executor-js/runtime-workerd-subprocess
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @executor-js/codemode-core@1.5.33
+
 ## 0.0.4
 
 ### Patch Changes