+++ packages/plugins/apps/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @executor-js/plugin-apps
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @executor-js/sdk@1.5.33
+  - @executor-js/api@1.4.53
+  - @executor-js/react@1.4.53
+  - @executor-js/runtime-workerd-subprocess@0.0.5
+
 ## 0.1.3
 
 ### Patch Changes