+++ packages/plugins/encrypted-secrets/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @executor-js/plugin-encrypted-secrets
 
+## 0.0.32
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @executor-js/sdk@1.5.33
+
 ## 0.0.31
 
 ### Patch Changes